+++ integration_test/protocol_versions.js
@@ -1,6 +1,7 @@
 var WebSocketServer = require('ws').Server,
     Leap = require('../lib');
 
+
 setTimeout(function() {console.log('Failure, timing out.'); process.exit(1);}, 30000);
 
 downgradeProtocol = function(){
@@ -10,7 +11,7 @@ downgradeProtocol = function(){
   // Once a controller tried to connect, it will always be trying to reconnect. For now we just use two ports.
   var passed = false;
 
-  var expected = ['/v4.json', '/v3.json', '/v2.json', '/v1.json'];
+  var expected = ['/v5.json', '/v4.json', '/v3.json', '/v2.json', '/v1.json'];
 
   var wss = new WebSocketServer({port: 9494})
   wss.on('connection', function(ws) {
@@ -25,7 +26,9 @@ downgradeProtocol = function(){
       }
     } else if (ws.upgradeReq.url == expected[0]) {
       expected.shift();
-      ws.close(1001);
+      // for some reason, the response gets eaten without this.
+      setTimeout(function(){ws.close(1001);}, 100)
+//      ws.close(1001);
     } else {
       console.log("FAILED downgradeProtocol, expected: "+JSON.stringify(expected[0])+" , got ws.upgradeReq.url:"+ws.upgradeReq.url);
       process.exit(1);